Numerous bigfixes and player shooting improvement
We fixed most bugs for zombies not detecting the player and structures properly due to new tree and player collider objects that were obscuring their linecasts. Improved the player shooting mechanic by turning the player according to the position of where he clicked with the mouse to shoot on the screen. Changed the button for the pause menu and updated the background in the end-game screen with the new dark splash background